refactor(core): move GitHub-settings + repo-hygiene checks into src/base (#282)
The branch-protection / merge-settings / workflow-permissions checks
(#137/#138) and the repo-hygiene checks are language-agnostic — every
language module gets them for free once they live in src/base.

- Move github-settings.ts (checks + apply logic) to src/base/, flipping
  its CheckResult import from ../commands/doctor.js to ./types.js
  (removes a base->cli back-reference). Test moves to tests/base/.
- Move 6 pure-dir hygiene checks into src/base/checks.ts:
  checkGitHubActions, checkCodeQL, checkDependabot, checkGitLabCI,
  checkCoverageUpload, checkAiSetup.
- Update the two github-settings importers (doctor.ts, js/fixers.ts).

Left in doctor.ts: checkReleaseToken + checkNpmOidcPublish
(semantic-release/npm — JS-ecosystem) and checkLockfile (its Lockfile
type is cli-coupled, so moving it would be a backward layer dep). These
await per-module dispatch (#285).

Pure refactor, zero behaviour change: byte-identical doctor --json
snapshot over a 43-check fixture, typecheck + biome clean, 480 tests
green.

Closes #282
